Fix compilation with Boost > 1.47.
The singleton_pool.hpp is supposed to be header-only (and is document as
such).  Unfortunately, Boost 1.48 introduced a dependency on thread
library, which breaks compilation.

It's not desirable to make libxmlwrapp depend on boost libraries, so
just disable the use of boost::singleton_pool in this case.  Hopefully
Boost folks will fix this one eventually.

See e.g.  here for reference:
https://svn.boost.org/trac/boost/ticket/7335
http://stackoverflow.com/questions/10447424/using-boost-pool-in-vs2010-linker-error/10447677