Add xml::document ctors for parsing.
Avoid the need to use tree_parser class, add convenience constructors to
create xml::document from data or from file.  This is consistent with
xml::schema and xslt::stylesheet classes.

Note that the document(filename) ctor cannot have default value of
on_error argument because of ambiguity with the existing
document(rootNodeName) ctor.  Which is why the latter is deprecated.