Make dumping xml::nodes more convenient.
Add an overload of node_to_string() just returning the string with the entire
node contents instead of requiring to pass in a string to fill.

The main benefit of this change is that the new function can be called from a
debugger to view the node contents during a debugging session.

Also mention the existence of this method in the "Tips and tricks" section of
the manual.