Fix check for boost::singleton_pool in configure
This check could never pass with a standard conforming C++03 compiler (such as
g++ 4.x) because it uses a locally defined (inside the function main) struct
tag as a template argument, so compiling the test program always failed.

Confusingly, it did succeed in C++11 mode which resulted in other problems in
this case but not when using C++98 even though they didn't have anything to do
with the language dialect choice.