Fix crashes in some of xslt::stylesheet::apply() overloads
If the provided on_error handler doesn't throw an exception (e.g. if it's an
xml::error_messages object), don't crash in apply() overloads returning the
result of the transformation.

Just check for apply_stylesheet() success there too, just as it was already
done in the overloads using the "result" document object.