Allow skipping building a single test case using zlib
This is especially useful under Windows where zlib might not be available out
of the box, unlike under Unix systems which always have it, because the test
suite can avoid linking to it then.

By default, zlib is still required and XMLWRAPP_NO_ZLIB needs to be explicitly
predefined to avoid depending on it.