Don't leak memory when creating a document object fails
There is no need to allocate a new doc_impl when creating a document from file
or data, it's not used at all in the normal case after a call to swap() and is
simply leaked if an error occurs.

This fixes memory leak reports from MSVC debug CRT and, perhaps more
importantly, avoids an unnecessary allocation in the normal case.