Improve error handling in document::save_to_{string,file}()
Throw an exception if any errors occur while saving (this can happen if the
document contains nodes with not well-formed UTF-8 text, for example) instead
of ignoring any errors and silently losing data.

Also make it possible to pass a custom on_error handler to these functions,
just as it was already possible when loading the document.