Don't suppress all libxml2 errors by default
It's not especially useful to show libxml2 error messages on stderr,
especially under Windows, but it's even worse to not show them anywhere at all
and silently ignore all errors.

Don't do this any longer, errors should be handled by adding code using
xml::impl::global_errors_collector to the functions where they may happen, as
it was done in the previous commits on this branch.