Return libxml2 IO errors correctly to the user code
It's simpler and better (as libxml2 may have more details) to just return the
errors logged by libxml2 global error handler to the caller, instead of
checking whether the input file could be opened ourselves.

This does change the format of the error message, which now takes the form

	I/O warning : failed to load external entity "$filename"

so update the unit test to make it less strict: we don't really care about the
exact message, just that it mentions the file that we failed to open.