Allow to optionally throw exceptions for invalid XPath expressions
To preserve compatibility, continue to ignore errors when evaluating XPath, as
the caller may not care about them and the empty range as the return value
already indicates that an error occurred.

However also allow throwing an exception on error, for consistency with the
other methods taking an "error_handler".