Improve libxml2 error reporting by including more details
Previous attempts to report errors better were insufficient as the generic
error handler in libxml2 is used multiple times when reporting errors by
libxml2 and as we threw after getting the first one, it resulted in such error
message as "I/O " (and nothing else) when failing to load a file.

Fix this by setting up the structured error handler which is called atomically
and also provides more information than what is passed to the generic handler.
Notice that we still need to continue using generic error handler as some
places in libxml2 only use it and not the structured handler.

Add a unit test verifying that the name of the file which couldn't be opened
appears in the error message of the exception thrown in this case.