Always use UTF-8 encoding by default in xml::document
This was already mostly the case, i.e. UTF-8 was (implicitly) used if
set_encoding() had been never called, but the default encoding was
change to ISO-8859-1 if get_encoding() was called, even without any
preceding set_encoding() call, which didn't make any sense at all.

So stop changing the default encoding in get_encoding() and just return
"UTF-8" from it if no encoding had been previously set.

This change also allows us to get rid of unnecessary doc_impl::encoding_
variable which, other than being used in the get_encoding() hack
described above, seemed to exist solely in order to allow returning a
reference to it from document::get_encoding(), but this seems
unnecessary when we can just return the string by value (which is mostly
backwards compatible with the existing code) and removing this variable
simplifies the code as we don't need to keep it in sync with the
encoding stored in xmlDocPtr any more.

Also stop using xmlDocDumpFormatMemoryEnc() and xmlSaveFormatFileEnc()
when we can use the simpler functions without the "Enc" suffix doing the
same thing anyhow.