Remove the need to specify --foreign option with automake
Add "foreign" to AM_INIT_AUTOMAKE() macro used in configure instead to
ensure that automake always runs in "foreign" mode (in which it doesn't
require such GNU-specific files as COPYING or INSTALL), even if it's not
specified on its command line.

This allows running "autoreconf -I admin" to regenerate the
autotools-related parts of the build system instead of running the
booststrap script which also runs doxygen and bakefile, even if it's not
really necessary.