Remove the need to specify "-I admin" on aclocal command line
Similarly to the previous commit, record the location of the extra
include directory containing custom macros in configure.ac itself, so
that aclocal gets it from there, even if it's not specified on its
command line.

This allows to run "autoreconf" without any options to regenerate
configure and Makefile.in.

Notice that AC_CONFIG_MACRO_DIR() is deprecated now in favour of
AC_CONFIG_MACRO_DIRS(), but the latter requires a relatively recent
version of automake (1.13) and there doesn't seem to be any benefit in
requiring it now, the macro can always be trivially replaced with the
"S" version when autoconf starts warning about it.