Stop using Boost.Pool library
Benchmarks show that not only there are no noticeable benefits from
using it but that, actually, the net effect is almost always negative,
albeit slightly so as in any realistic benchmark scenario the run-time
is completely dominated by libxml2 functions.

Remove this optional dependency as the first step towards removing all
dependencies on Boost from the library.