Clean up redundant namespace definitions when inserting new nodes
Copying a node with an implicitly defined (i.e. inherited from its
parent) namespace resulted in creating a namespace definition for the
node, even if its implicitly inherited namespace in the new location
would have been exactly the same.

While not exactly wrong, this was ugly and made updating the namespace
for all nodes in a document with the only namespace defined at the
document root level more difficult as it had to be done recursively, for
all nodes, instead of just changing the namespace of the root.

Fix this by removing the namespace definition from the node created by
xmlCopyNode() if it's the same as the namespace the new node would have
used anyhow.

Add a unit test checking that there is no redundant namespace definition
in the new node.