Check for the node being valid when erasing it
Throw an exception rather than crash if an attempt to delete an invalid
node is made using any of the public erase() methods (in node,
nodes_view or document classes).

Note that this also fixes gcc -Wnull-dereference warning given when
compiling the library as an amalgamation, i.e. as a single translation
unit containing all of its source files together, as a side effect.