Check for overflow when converting size_t to int
This fixes a few -Wconversion gcc warnings and, more importantly,
actually ensures that there are no integer overflows in the code which
might result in real problems.

Add checked_int_cast() helper which is used to check for such overflows
in all the places where gcc flags of them and throws an exception
instead of using a wrong (wrapped) value if it happens.