Enable more useful warnings not part of -Wall by default
In combination with the use of -Werror for the CI builds, this will
catch any warnings introduced in the new code, while not being too
bothersome for people compiling the library without -Werror, i.e. most
of the normal users.

Note that of these options are not supported until gcc 6, so we need to
check whether the compiler being used has them before appending them to
CXXFLAGS. A helper macro from

https://www.gnu.org/software/autoconf-archive/ax_append_compile_flags.html

is used to do this easily.