Remove unnecessary initxmlDefaultSAXHandler definition
This source file doesn't use initxmlDefaultSAXHandler nor any other
symbols from SAX2.h, so don't include it at all, as this is simply
unnecessary here.

Incidentally, this also results in problems when using libxml2 2.9.9 or
later and "unity" builds of xmlwrapp, in which all of the source files
are compiled as a single translation unit, because the definition here
breaks SAX.h inclusion in src/libxml/tree_parser.cxx when it's included
later.