Replace hack with initxmlDefaultSAXHandler() redefinition
This was ugly and dangerous and completely unnecessary. Just define our
own xmlwrapp_initDefaultSAXHandler() instead of redefining libxml2
symbol and potentially breaking its headers inclusions -- as it happened
in the "unity" builds of xmlwrapp until the parent commit.

Of course, we could just drop compatibility with libxml2 < 2.6.0 (which
is 17 years old by now), but this can/will be done later.