Fix BSD 3-Clause license wording to match official
The BSD license uses "copyright holder" term, not "author", and we should
do the same both because it is better terminology, and because we don't
want any spurious divergences from the standard license.

The terms of the license are not changed by this.

This improves detection of the license by automated tools.

Fixes #51