Remove bespoke bootstrap script and just use autoreconf instead
The script is not really necessary and can actually create problems, for
example when running it on the systems with doxygen installed but
without dot, as it then generates a ton of warning messages about the
latter being unavailable, which is surprising and also wasteful if we
don't want to generate the docs in the first place (as bootstrap has to
be run before running "./configure --disable-docs" too).

So simplify and streamline things by just getting rid of bootstrap
script entirely.