Change node::set_namespace() to always update children namespaces
This function already updated the children namespaces when changing the
existing namespace, but not when setting a new namespace for a node
which hadn't previously had any.

This was unexpected and resulted in a tree with broken namespace
inheritance invariant and, as the result of this, serializing such a
tree to string and deserializing it back from it resulted in a different
tree, with correct namespace inheritance.

Fix this by always maintaining the invariant, even if it requires
iterating over all node children, i.e. can't be done in constant time in
this case.

Add a unit test checking for the new behaviour and document it in the
changelog.