Use xmlSetNs() instead of setting ns pointer directly
This improves encapsulation and also avoids setting this pointer for the
tree nodes for which it doesn't make sense to use it, i.e. anything but
elements and attributes -- even though this didn't seem to have any ill
effects, it's still better to avoid setting namespace on "text" nodes,
for example.