Build inside container
Mostly do it in order to be able to upload the entire /usr/local which
is empty inside the minimal container but contains plenty of other stuff
in the default GitHub Actions environment -- this is much simpler than
uploading just the files installed by "make install" or installing them
in some other directory.

Also separate 32- and 64-bit builds.