Work around libtool silently ignoring LDFLAGS
There is a libtool problem known since at least 15 years, see e.g. [1],
and regularly being rediscovered again and again by various people (see
e.g. [2]), which still remains without any solution other than patching
libtool itself: it simply ignores "unknown" flags in LDFLAGS, including
the very useful -static-libgcc one, which we use to avoid dependencies
on gcc run-time DLLs here.

The only solution seems to be to pass the flag explicitly on libtool
command line, as it does, at least, pass through -Wl,xxx or, to make it
clear that this is a libtool flag, -Wc,xxx, but this runs into bugs in
libxslt configure which tried to (incorrectly) work around another
long-standing libtool problem with having to specify -no-undefined
explicitly when building DLLs, even if it's always required there.

Work around that workaround by including this option in the flags we
pass to libtool and, hopefully, this will work for some time. And
hopefully in the long term we (and the world) are going to stop using
libtool.

[1]: https://www.mail-archive.com/cygwin@cygwin.com/msg99768.html
[2]: https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/7BRVA3N6C7GCWLY6LT72MPNSIHXBMBQ2/