Replace old static library MSVS projects with new DLL ones
Drop support for non MSBuild-based MSVS versions (< 2010) as they are
obsolete and soon won't be supported for building the code anyhow as
C++14 will be required.

Add new bkl-generated projects to VCS: this goes against the general
rule of not storing generated classes in it, but is too convenient in
practice to not do it and doesn't add that many files now that we can
use bkl "msvs" toolset to reuse the same files for all MSVS versions.

Also switch to using DLLs instead of static libraries for xmlwrapp
itself, as this is more useful in practice and also more likely to
result in problems, so this is what we want to test in the CI builds.

Finally, make the location of libxml2 and friends more flexible by
extracting the related options into libxml.props file which can be
edited if the libraries are available in a different location or even
removed entirely if they're installed globally.