Avoid MSVS warning about inheriting from non DLL-exported class
DLL-export all the public function of xslt::init instead of exporting
the entire class itself, as this seems to be the least annoying way to
avoid MSVS warning C4275 about using a non DLL-exported (from XSLT DLL)
xml::init class as a base class for the DLL-exported xslt::init.

This shouldn't result in any real changes.