Set ELF visibility of xml::exception to public
This is important for clang under macOS, as otherwise exceptions of this
type are not caught properly in the applications using xmlwrapp.

Doing this required adding yet another macro, as just using the existing
XMLWRAPP_API for this class would have resulted in warnings from MSVS
about exporting a class inheriting from a non-DLL-exported standard one.