Use Ubuntu 16.04 instead of CentOS 7 for a CI build with old gcc
We want to test with old gcc, but 4.8 in CentOS 7 is too old and
installing slightly newer compiler there is too difficult, as 7.9 only
has devtoolset-4 or later, i.e. gcc-7, and while 7.6 still has
devtoolset-4, setting it up to use it is not that simple, so just use
Ubuntu 16.04 which has gcc-5 instead -- this should be old enough in
practice.