Disable MSVS DLL interface member warnings more globally
Using the pair of macros around every member of the standard type is too
annoying and ugly, so put them around the entire header contents for all
files that need them.

It's still slightly ugly, but is better than disabling the warning
globally in case the user code cares about it.