Update to Catch v2
Replace the existing CATCH v1 header with a single header version of v2:
this is not how it's supposed to be used (it should normally be included
as catch2/catch.hpp), but as long as we always use the internal header
and not the system provided one it doesn't really matter.

The only changes needed in the tests were to remove "&" from the
exception specification passed to {CHECK,REQUIRE}_THROWS_AS as they need
the exception type name, not the full type, and using "&" in it doesn't
compile with v2.

No real changes.

See #86.