Build either static or shared libraries, not both at once
Behave in a standard (for CMake projects) way and build either static or
shared libraries depending on whether BUILD_SHARED_LIBS is on or off.

Also use libxml2 as PUBLIC library dependency of xmlwrapp to ensure that
all targets depending on xmlwrapp link with libxml2 as well instead of
explicitly specifying libxml2 libraries for them too.

And do the same for libxslt and libexslt and xsltwrapp too.