Move config file generation to installation time
This is required to honour the prefix chosen for the installation,
otherwise the prefix used at configuration time was always used, but it
may be different from the actually used one when using a command like

	cmake --install builddir --prefix /some/where

This also avoids the needs for an extra function -- but now requires an
extra file (in principle, all its code could be inlined, but this would
be too ugly).

As a side effect, increase the minimum required version to 3.20. If this
is ever a problem, we can always stop using FILE_PERMISSIONS and use
file(CHMOD) instead.