Fix search for pkg-confg modules
Fix syntax of the version check: previously, we searched for any version
of libxml-2.0 and then, if not found, for ">=" and "2.4.28" modules (!).
Remove spaces to do what we really need to do.

Also use pkg_check_modules() rather than pkg_search_module() because the
latter is meant for searching for multiple modules, but here we know
which module we need, so use the simpler command instead.